+++ engine/src/main/java/me/molybdenum/ambience_mini/engine/client/configuration/abstract_syntax/expression/BinaryOperators.java
@@ -3,5 +3,6 @@ package me.molybdenum.ambience_mini.engine.client.configuration.abstract_syntax.
 public enum BinaryOperators {
     EQ, APP_EQ, MATCH,
     AND, OR, LT, LE,
-    INDEXER
+    INDEXER,
+    ADD, SUB, MUL, DIV
 }