+++ engine/src/main/java/me/molybdenum/ambience_mini/engine/client/configuration/abstract_syntax/type/CombatantT.java
@@ -1,0 +1,33 @@
+package me.molybdenum.ambience_mini.engine.client.configuration.abstract_syntax.type;
+
+import me.molybdenum.ambience_mini.engine.client.configuration.abstract_syntax.type.kinds.AccessibleT;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Map;
+
+public final class CombatantT extends Type implements AccessibleT
+{
+    public static final CombatantT INSTANCE = new CombatantT();
+
+    public static final String FIELD_TYPE_ID = "type_id";
+    public static final String FIELD_HEALTH = "health";
+    public static final String FIELD_MAX_HEALTH = "max_health";
+    public static final String FIELD_HEALTH_PERCENT = "health_percent";
+    public static final Map<String, Type> FIELDS = Map.of(
+            FIELD_TYPE_ID, StringT.INSTANCE,
+            FIELD_HEALTH, FloatT.INSTANCE,
+            FIELD_MAX_HEALTH, FloatT.INSTANCE,
+            FIELD_HEALTH_PERCENT, FloatT.INSTANCE
+    );
+
+
+    @Override
+    protected boolean equalToInternal(@NotNull Type other) {
+        return other instanceof CombatantT;
+    }
+
+    @Override
+    public Map<String, Type> fieldTypes() {
+        return FIELDS;
+    }
+}