+++ engine/src/main/java/me/molybdenum/ambience_mini/engine/client/configuration/interpreter/values/FloatVal.java
@@ -1,13 +1,13 @@
 package me.molybdenum.ambience_mini.engine.client.configuration.interpreter.values;
 
+import me.molybdenum.ambience_mini.engine.client.configuration.interpreter.values.kinds.AccessibleV;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Objects;
 
-public final class FloatVal extends Value<Float>
+public final class FloatVal extends Value<Float> implements AccessibleV
 {
     public static final FloatVal UNDEFINED = new FloatVal();
-    public static final FloatVal ZERO = new FloatVal(0f);
 
 
     public FloatVal() {
@@ -27,5 +27,10 @@ public final class FloatVal extends Value<Float>
     @Override
     public boolean equals(Value<?> other) {
         return Objects.equals(value, other.value);
+    }
+
+    @Override
+    public Value<?> getField(String field) {
+        return "intVal".equals(field) ? new IntVal(value.intValue()) : IntVal.UNDEFINED;
     }
 }