+++ engine/src/main/java/me/molybdenum/ambience_mini/engine/client/configuration/interpreter/values/IntVal.java
@@ -1,13 +1,13 @@
 package me.molybdenum.ambience_mini.engine.client.configuration.interpreter.values;
 
+import me.molybdenum.ambience_mini.engine.client.configuration.interpreter.values.kinds.AccessibleV;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Objects;
 
-public final class IntVal extends Value<Integer>
+public final class IntVal extends Value<Integer> implements AccessibleV
 {
     public static final IntVal UNDEFINED = new IntVal();
-    public static final IntVal ZERO = new IntVal(0);
 
 
     public IntVal() {
@@ -27,5 +27,10 @@ public final class IntVal extends Value<Integer>
     @Override
     public boolean equals(Value<?> other) {
         return Objects.equals(value, other.value);
+    }
+
+    @Override
+    public Value<?> getField(String field) {
+        return "floatVal".equals(field) ? new FloatVal(value.floatValue()) : FloatVal.UNDEFINED;
     }
 }