+++ engine/src/main/java/me/molybdenum/ambience_mini/engine/client/core/providers/GameStateProviderTemplate.java
@@ -39,6 +39,8 @@ public abstract class GameStateProviderTemplate extends BaseGameStateProvider
     public static final EventTemplateV1 E_FISHING = new EventTemplateV1("fishing", instance -> instance::isFishing);
     public static final EventTemplateV1 E_UNDER_WATER = new EventTemplateV1("under_water", instance -> instance::isUnderWater);
     public static final EventTemplateV1 E_IN_LAVA = new EventTemplateV1("in_lava", instance -> instance::inLava);
+    public static final EventTemplateV1 E_ON_FIRE = new EventTemplateV1("on_fire", instance -> instance::onFire);
+    public static final EventTemplateV1 E_IN_POWDER_SNOW = new EventTemplateV1("in_powder_snow", instance -> instance::inPowderSnow);
     public static final EventTemplateV1 E_DROWNING = new EventTemplateV1("drowning", instance -> instance::isDrowning);
 
     // Mount events
@@ -59,7 +61,7 @@ public abstract class GameStateProviderTemplate extends BaseGameStateProvider
             E_DAY, E_DAWN, E_DUSK, E_NIGHT,
             E_DOWNFALL, E_RAIN, E_SNOW, E_THUNDERING,
             E_VILLAGE, E_RANCH, E_WARDEN_NEARBY,
-            E_DEAD, E_SLEEPING, E_FISHING, E_UNDER_WATER, E_IN_LAVA, E_DROWNING,
+            E_DEAD, E_SLEEPING, E_FISHING, E_UNDER_WATER, E_IN_LAVA, E_ON_FIRE, E_IN_POWDER_SNOW, E_DROWNING,
             E_MINECART, E_BOAT, E_HORSE, E_DONKEY, E_PIG, E_ELYTRA,
             E_IN_COMBAT, E_BOSS_FIGHT
     };
@@ -89,7 +91,7 @@ public abstract class GameStateProviderTemplate extends BaseGameStateProvider
 
     // Combat properties
     public static final PropertyTemplateV1 P_COMBATANT_COUNT = new PropertyTemplateV1("combatant_count", new IntT(), instance -> instance::countCombatants);
-    public static final PropertyTemplateV1 P_COMBATANT_IDS = new PropertyTemplateV1("combatant_ids", new ListT(new StringT()), instance -> instance::getCombatantIDs);
+    public static final PropertyTemplateV1 P_COMBATANTS = new PropertyTemplateV1("combatants", new ListT(new CombatantT()), instance -> instance::getCombatants);
     public static final PropertyTemplateV1 P_BOSS = new PropertyTemplateV1("boss", new StringT(), instance -> instance::getBoss);
     public static final PropertyTemplateV1 P_BOSSES = new PropertyTemplateV1("bosses", new ListT(new StringT()), instance -> instance::getBosses);
 
@@ -100,7 +102,7 @@ public abstract class GameStateProviderTemplate extends BaseGameStateProvider
             P_DIFFICULTY, P_DIMENSION, P_BIOME, P_BIOME_TAGS, P_TIME, P_CAVE_SCORE, P_SKYLIGHT_SCORE,
             P_AREAS, P_STRUCTURES,
             P_UUID, P_GAME_MODE, P_HEALTH, P_MAX_HEALTH, P_ELEVATION, P_VEHICLE, P_EFFECTS,
-            P_COMBATANT_COUNT, P_COMBATANT_IDS, P_BOSS, P_BOSSES,
+            P_COMBATANT_COUNT, P_COMBATANTS, P_BOSS, P_BOSSES,
             P_FLAGS
     };
 
@@ -157,6 +159,8 @@ public abstract class GameStateProviderTemplate extends BaseGameStateProvider
     public abstract BoolVal isFishing();
     public abstract BoolVal isUnderWater();
     public abstract BoolVal inLava();
+    public abstract BoolVal onFire();
+    public abstract BoolVal inPowderSnow();
     public abstract BoolVal isDrowning();
 
 
@@ -208,7 +212,7 @@ public abstract class GameStateProviderTemplate extends BaseGameStateProvider
     // ------------------------------------------------------------------------------------------------
     // Combat properties
     public abstract IntVal countCombatants();
-    public abstract ListVal getCombatantIDs();
+    public abstract ListVal getCombatants();
     public abstract StringVal getBoss();
     public abstract ListVal getBosses();