+++ engine/src/main/java/me/molybdenum/ambience_mini/engine/client/core/state/BaseCombatState.java
@@ -1,5 +1,7 @@
 package me.molybdenum.ambience_mini.engine.client.core.state;
 
+import me.molybdenum.ambience_mini.engine.client.configuration.interpreter.values.CombatantVal;
+import me.molybdenum.ambience_mini.engine.client.configuration.interpreter.values.ListVal;
 import me.molybdenum.ambience_mini.engine.client.core.BaseClientCore;
 import me.molybdenum.ambience_mini.engine.client.core.setup.ServerSetup;
 import me.molybdenum.ambience_mini.engine.shared.utils.versions.AmVersion;
@@ -46,7 +48,10 @@ public abstract class BaseCombatState<TEntity, TVec3>
     // Abstract API
     public abstract boolean isEntityDead(TEntity entity);
     public abstract int getEntityId(TEntity entity);
+
     public abstract String getEntityResourceLocation(TEntity entity);
+    public abstract Float getEntityHealth(TEntity entity);
+    public abstract Float getEntityMaxHealth(TEntity entity);
 
     public abstract boolean inBossFight();
     public abstract List<String> getBosses();
@@ -61,12 +66,6 @@ public abstract class BaseCombatState<TEntity, TVec3>
         if (hasActiveCombatants())
             return combatants.size(); // If in creative, there is no combat
         return 0;
-    }
-
-    public List<String> getCombatantResourceLocations() {
-        if (hasActiveCombatants())
-            return combatants.values().stream().map(c -> getEntityResourceLocation(c.entity)).toList();
-        return List.of();
     }
 
     public boolean hasActiveCombatants() {
@@ -125,6 +124,10 @@ public abstract class BaseCombatState<TEntity, TVec3>
         combatants.clear();
     }
 
+    public ListVal getCombatants() {
+        return new ListVal(combatants.values().stream().map(Combatant::asCombatantVal));
+    }
+
 
     class Combatant {
         public TEntity entity;
@@ -132,6 +135,14 @@ public abstract class BaseCombatState<TEntity, TVec3>
 
         public Combatant(TEntity entity) {
             this.entity = entity;
+        }
+
+        public CombatantVal asCombatantVal() {
+            return new CombatantVal(
+                    getEntityResourceLocation(entity),
+                    getEntityHealth(entity),
+                    getEntityMaxHealth(entity)
+            );
         }
     }
 }