+++ v1_18_2/src/main/java/me/molybdenum/ambience_mini/v1_18_2/client/core/state/CombatState.java
@@ -8,6 +8,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.fml.util.ObfuscationReflectionHelper;
 import net.minecraftforge.registries.ForgeRegistries;
@@ -38,6 +39,16 @@ public class CombatState extends BaseCombatState<Entity, Vec3>
     public String getEntityResourceLocation(Entity entity) {
         var key = ForgeRegistries.ENTITIES.getKey(entity.getType());
         return key == null ? null : key.toString();
+    }
+
+    @Override
+    public Float getEntityHealth(Entity entity) {
+        return (entity instanceof LivingEntity liv) ? liv.getHealth() : null;
+    }
+
+    @Override
+    public Float getEntityMaxHealth(Entity entity) {
+        return (entity instanceof LivingEntity liv) ? liv.getMaxHealth() : null;
     }